\name{gausspolytopewalk-package}
\alias{gausspolytopewalk-package}
\alias{gausspolytopewalk}
\docType{package}
\title{
\packageTitle{gausspolytopewalk}
}
\description{
\packageDescription{gausspolytopewalk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gausspolytopewalk}
\packageIndices{gausspolytopewalk}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gausspolytopewalk}

Maintainer: \packageMaintainer{gausspolytopewalk}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
